Fix HTML entities in browse button title
With plone.protect auto protection enabled all non-ascii characters in
the output get replaced by their HTML entities. But when setting an
attribute with the jQuery attr method HTML entities are not resolved.
See http://bugs.jquery.com/ticket/6780 for more information about this
issue.

This commit works around this by changeing the JS code to not use the
attr method the set the button title.